Create BrodyStringerNFTs
I am a YouTuber who is looking to distribute a channel-related fungible token to my first 1000 subscribers. I have been distributing them individually from my nami wallet and it is quite tedious. I have also spent all the Cardano I had set aside for fees for the giveaway and now need the purchaser to put forward just enough ada to cover the fees. I hope to sell my nft at the lowest price possible as to allow the most viewers to get a hold of one.

Here is a link to my youtube channel: 
https://www.youtube.com/channel/UCouAchDoFe47kpj9MRN0rFg 

As of making this video I have 557 Subscribers and 1 NFT to distribute, but with continued growth in NFT usability, and Cardano as a whole I plan on making additional NFT's.
I speak in many of my videos about doing PizzaMoneyNFT Give Aways and I now wish to list those Pizza Money NFT's.

Cheers,
Brody Stringer

PS: The NFT in question has a water mark in the bottom right corner that says brodystringerNFT's for further authentication.